# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-Commerce",
        "description": "Documentación de los endpoints del E-Commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Obtener órdenes del usuario autenticado",
                "description": "Devuelve una lista paginada de las órdenes pertenecientes al usuario autenticado, incluyendo sus ítems y pagos.",
                "operationId": "getUserOrders",
                "responses": {
                    "200": {
                        "description": "Lista paginada de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear una nueva orden",
                "description": "Crea una orden con sus ítems, verifica el stock disponible de los productos y descuenta las existencias dentro de una transacción.",
                "operationId": "storeOrder",
                "requestBody": {
                    "description": "Lista de productos y cantidades a pedir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación (ej. Stock insuficiente o producto inactivo)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuficiente para Camiseta Deportiva."
                                        },
                                        "errors": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Stock insuficiente para Camiseta Deportiva."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Obtener detalle de una orden",
                "description": "Muestra la información de una orden específica si pertenece al usuario autenticado.",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Acceso no autorizado a la orden"
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "description": "Modelo de Orden de compra",
                "required": [
                    "user_id",
                    "status",
                    "subtotal",
                    "total",
                    "currency"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "paid",
                            "cancelled",
                            "completed"
                        ]
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 49.98
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 49.98
                    },
                    "currency": {
                        "type": "string",
                        "example": "USD"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-09-13T20:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-09-13T20:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}